Add char counter to description field
* Show "N/1000" counter under "Description" field at all times
* Clean up a bunch of styles – use Bootstrap's form classes, remove
  custom styles

It would be nice to only show the char counter when the field is
actually highlighted, but I didn't feel too strongly about how that
should look or behave. If the whole thing is just removed from the page,
then the rest of the form bounces around as the description field is
highlighted, and that's ugly. If it's just made invisible, then there's
awkward spacing one way or another.